+++ package.json
@@ -29,7 +29,7 @@
     "moment": "^2.29.4",
     "react": "18.2.0",
     "react-native": "0.71.6",
-    "react-native-calendar-picker": "^7.1.3",
+    "react-native-calendar-picker": "^6.1.5",
     "react-native-dotenv": "^3.4.8",
     "react-native-navigation": "^7.32.1",
     "react-native-paper": "^5.7.0",
