+++ src/api/userApi.js
@@ -4,40 +4,6 @@ import {Alert} from 'react-native';
 
 let baseUrl = process.env.BASE_URL;
 
-/* export async function postRegister(
-  lastName,
-  firstName,
-  email,
-  password,
-  isBlind,
-  //gender,
-  navigation,
-) {
-  axios
-    .post(`${baseUrl}/auth/register`, {
-      lastname: lastName,
-      firstname: firstName,
-      email: email,
-      password: password,
-      isBlind: isBlind,
-      //gender: gender,
-    })
-    .then(async function (response) {
-      const tokenData = JSON.stringify(response.data.token);
-      await AsyncStorage.setItem('Token', tokenData);
-      if ((response.status = '200')) {
-        const getTokenData = await AsyncStorage.getItem('Token');
-        return getTokenData != null
-          ? JSON.parse(getTokenData) &&
-              navigation.navigate('FormRouteBlind', {token: getTokenData})
-          : null;
-      }
-    })
-    .catch(function (error) {
-      Alert.alert('erreur : ' + JSON.stringify(error));
-    });
-} */
-
 export async function axiosRegiter(
   firstName,
   lastName,
