+++ src/views/blind/formRouteBlind.js
@@ -18,6 +18,7 @@ import {
   TouchableOpacity,
 } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AxiosRoute } from '../../api/routeApi';
 
 const FormRouteBlind = ({route, navigation}) => {
   const routeParamsToken = route.params.token;
@@ -27,87 +28,14 @@ const FormRouteBlind = ({route, navigation}) => {
   const [time, setTime] = useState();
   const [routeMateGender, setRouteMateGender] = useState();
 
-  const postRoute = useCallback(async (fromStation, toStation) => {
-    axios.defaults.headers.common['Authorization'] = `Bearer ${JSON.parse(
-      routeParamsToken,
-    )}`;
-    console.log('routeParamsToken form 1: ' + JSON.parse(routeParamsToken));
-
-    axios
-      .post('http://localhost:8080/api/v1/routes/add', {
-        fromStation: fromStation,
-        toStation: toStation,
-        routeMateGender: routeMateGender
-      })
-      .then(async function (response) {
-        const getAsynTokenStorage = await AsyncStorage.getItem('Token');
-        console.log(
-          'routeParamsToken async:' + JSON.parse(getAsynTokenStorage),
-        );
-        if ((response.status = '200' && fromStation && toStation && routeMateGender)) {
-          const fromStationData = JSON.stringify(fromStation);
-          console.log('fromstationPost : ' + fromStationData);
-          const toStationData = JSON.stringify(toStation);
-          console.log('tostationDataPost : ' + toStationData);
-          const routeMateGender = JSON.stringify(routeMateGender);
-          console.log('tostationDataPost : ' + toStationData);
-          
-          return getAsynTokenStorage != null
-            ? JSON.parse(getAsynTokenStorage)
-            : null;
-        }
-      })
-
-      .catch(function (error) {
-        if (!fromStation || !toStation) {
-          console.log('champ vide');
-          alert('erreur : ' + JSON.stringify(error));
-          console.log('perdu ! ' + error);
-        }
-      });
-
-    // axios
-    //   .get('http://localhost:8080/api/v1/routes/matches', {
-    //     fromStation: fromStation,
-    //     toStation: toStation,
-    //     routeMateGender: routeMateGender
-    //   })
-    //   .then(
-    //     function (response) {
-    //       if (response.data) {
-    //         navigation.navigate('Waiting');
-    //         console.log('dans response data get');
-    //         axios
-    //           .post('http://localhost:8080/api/v1/sendgrid', {})
-    //           .then(async function (response) {
-    //             console.log('dans le post sendgrid');
-    //           });
-    //       }
-    //     }
-
-          // const getAsynTokenStorage = AsyncStorage.getItem('Token');
-          //  const fromStationData = fromStation;
-          //  console.log('fromstationGet : ' + fromStationData);
-          // const toStationData = toStation;
-          //  console.log('tostationDataaGet : ' + toStationData );
-          // return getAsynTokenStorage != null ? JSON.parse(getAsynTokenStorage) && navigation.navigate('Match', { fromStation: fromStationData, toStation: toStationData, token: getAsynTokenStorage }) : null;
-
-          // } else {
-          //   navigation.navigate('Waiting')
-          // }
-          //  console.log(response.data[0]['fromStation'] + response.data[0]['fromStation'] )
-
-  //         .catch(function (error) {
-  //           if (!response.data || response.data == null) {
-  //             console.log('get erreur reponse vide ou null !');
-  //           } else {
-  //             console.log('get erreur !');
-  //             alert('erreur : ' + JSON.stringify(error));
-  //             console.log('perdu ! ' + error);
-  //           }
-  //         }),
-  //     );
-   }, []);
+  const postRoute = useCallback( () => {
+      AxiosRoute(fromStation,
+        toStation,
+        routeMateGender,
+        routeParamsToken, navigation)}, [fromStation,
+          toStation,
+          routeMateGender,
+          routeParamsToken, navigation]);
 
   return (
     <SafeAreaView style={{marginTop: '50%'}}>
@@ -155,7 +83,11 @@ const FormRouteBlind = ({route, navigation}) => {
         />
       </View>
       <TouchableOpacity
-        onPress={() => postRoute(fromStation, toStation)}
+        onPress={() => postRoute(fromStation,
+          toStation,
+          routeMateGender,
+          routeParamsToken, 
+          navigation)}
         onLongPress={() => console.log('pas de match désolé')}>
         <Text>Valider</Text>
       </TouchableOpacity>
